Create new modules for denorm assets
