+++ src/pudl/etl/__init__.py
@@ -8,7 +8,6 @@ from dagster import (
     Definitions,
     define_asset_job,
     load_assets_from_modules,
-    load_assets_from_package_module,
 )
 
 import pudl
@@ -46,7 +45,8 @@ default_assets = (
     *load_assets_from_modules([pudl.transform.ferc714], group_name="clean_ferc714"),
     *load_assets_from_modules([glue_assets], group_name="glue"),
     *load_assets_from_modules([static_assets], group_name="static"),
-    *load_assets_from_package_module(pudl.output, group_name="outputs"),
+    *load_assets_from_modules([pudl.output.denorm_eia], group_name="denorm_eia"),
+    *load_assets_from_modules([pudl.output.denorm_ferc], group_name="denorm_ferc"),
 )
 
 default_resources = {
