+++ src/pudl/output/ferc1.py
@@ -5,7 +5,6 @@ import sqlalchemy as sa
 
 import pudl
 from pudl.metadata.fields import apply_pudl_dtypes
-from pudl.output.sql.helpers import sql_asset_factory
 
 logger = pudl.logging_helpers.get_logger(__name__)
 
@@ -35,11 +34,6 @@ def read_table_with_start_end_dates(
         .where(table.columns.report_year <= f"{end_date.year}")
     )
     return pd.read_sql(table_select, pudl_engine)
-
-
-denorm_plants_utils_ferc1_asset = sql_asset_factory(
-    "denorm_plants_utils_ferc1", {"plants_ferc1", "utilities_ferc1"}
-)
 
 
 def plants_utils_ferc1(pudl_engine):
