+++ src/pudl/output/sql/__init__.py
@@ -1,0 +1,1 @@
+"""A module of python helper functions and sql files for creating SQL views."""
